+++ src/main/kotlin/de/mirb/project/miftp/config/BuildInfo.kt
@@ -1,0 +1,3 @@
+package de.mirb.project.miftp.config
+
+data class BuildInfo(var version: String, var timestamp: String)
