+++ src/main/resources/application.properties
@@ -20,6 +20,8 @@ miftp.eventListener.failOnMissing=false
 # the parameters are read as a map
 miftp.eventListener.parameters={paraKeyOne: 'first', paraKeyTwo: 'second' }
 
+miftp.maxTokenFiles=40
+
 # Sample for the Slack Notifier
 #miftp.eventListener=SlackNotifier
 #miftp.eventListener.failOnMissing=true
